requirements.txt
psutil==5.8.0
GPUtil==1.4.0
ping3==4.0.3
speedtest-cli==2.1.3
